refactor(hooks): move to own folder
