+++ app/components/inputs/RadioGroup.tsx
@@ -1,6 +1,6 @@
 import { useField } from "@rvf/react-router";
 import { type ReactNode, useState } from "react";
-import { useJsAvailable } from "~/services/useJsAvailable";
+import { useJsAvailable } from "~/services/hooks/useJsAvailable";
 import { isFieldEmptyOrUndefined } from "~/util/isFieldEmptyOrUndefined";
 import { type ErrorMessageProps } from ".";
 import InputError from "./InputError";
