+++ app/services/hooks/useJsAvailable.ts
