+++ app/services/hooks/useShouldPrint.ts
